Refactor: Move TvSeriesMapperTest to its own file
